Redesign of event recording and cell updating.
Each cell now tracks what events it is dependent upon.
Each event now tracks what cells are dependent upon it.
This information is kept up to date by the recorder.

We can use this information to recursively calculate all
cells that would need to be updated if a given event were
to fire.

Ultimately, all reactivity has always been handled by DepCells,
and by their ability to record dependencies and refresh. This
has not changed. It is, however, more explicit. Each ObsCell
now subscribes to its onSet event when it is created. Whenever
its onSet event fires, it calculates all of its downstream
dependencies and topographically sorts them.*

Naively, this approach would lead to all descendants being
calculated and refreshed at each level of the tree--which
would be unfortunate. To avoid this problem, when a cell
begins refreshing, it and all of its downstream cells
are marked as shielded, which means that they will
ignore any further refresh calls. This is so that they
can publish onSet and events without triggering additional
refreshes downstream.

This is also useful for debugging, as it is now trivially
simple to calculate the upstream and downstream dependencies
of any given cell.

In addition, transactions now track the EVENTS that they
will be publishing, not merely the listeners that they will
be calling. This means that, similar to the new cell refreshing
pattern above, we can calculate the topographically ordered
set of all cells that will be refreshed as the result of the
transaction's events and refresh them, after which we fire off
all of the saved events. Again, we make use of the shielding
mechanism to prevent duplicate updates and DAG recalculation.

Finally, these changes have resulted in some simplifications.

First, DepMgr no longer records events subscriptions,
since those records were not actually being used anywhere.
It is now solely used for managing transactions.

Second, recorder.sub is now considerably less flexible.
Previously, it had only ever been used to track cell refreshes,
but had the flexibility to do many other things. This, however,
led to some unnecessarily confusing code. Now, recorder.sub
only handles dependency tracking and cell refreshing.

Third, fixed a bug where, if an event listener is called
in a transaction when the listener itself spawns another transaction,
an infinite loop would result.

* To sort our dependencies, we traverse the dependency DAG in
  breadth first order. At each level of the traversal, we deduplicate
  the currently reached nodes, and then traverse to their children,
  until the DAG has been exhausted. We then flatten and reverse this
  list of nodes, and deduplicate it one final time.