Non-templating portions now handled by bobtail-rx.
Keeping the DOM DSL tightly coupled to the FRP
primitives has always been a tad problematic.
We have now separated these primitives into their
own package, so that they can be imported without
the DSL. This is useful for those using bobtail
in a non-browser environment such as Node.

The bobtail package continues to expose those
reactive programming primitives, with the DSL
attached to the rx.rxt namespace, as before.